+++ mmms-backend/src/main/java/com/sep/mmms_backend/controller/CommitteeController.java
@@ -1,16 +1,16 @@
 package com.sep.mmms_backend.controller;
 
 
+import com.sep.mmms_backend.dto.CommitteeDetailsDto;
 import com.sep.mmms_backend.entity.Committee;
 import com.sep.mmms_backend.response.Response;
 import com.sep.mmms_backend.response.ResponseMessages;
 import com.sep.mmms_backend.service.CommitteeService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/api")
@@ -52,5 +52,28 @@ public class CommitteeController {
     public ResponseEntity<Response> createCommittee(@RequestBody Committee committee, Authentication authentication) {
         committeeService.createCommittee(committee, authentication.getName());
         return ResponseEntity.ok().body(new Response(ResponseMessages.COMMITTEE_CREATION_SUCCESS));
+
+    }
+
+    @GetMapping("/getCommittees")
+    public ResponseEntity<Response> getCommittees(Authentication authentication) {
+        List<Committee> committees =  committeeService.getCommittees(authentication.getName());
+        return ResponseEntity.ok().body(new Response(committees));
+    }
+
+
+
+
+    /**
+     * this route fetches the committee from the database, checks if the committee is accessible by the current user
+     *
+     * then fetches all the members associated with the committee and sends both of them as response
+     */
+
+
+    @GetMapping("/getCommitteeDetails")
+    public ResponseEntity<Response> getCommitteeDetails(@RequestParam int committeeId, Authentication authentication) {
+        CommitteeDetailsDto committeeDetails = committeeService.getCommitteeDetails(committeeId, authentication.getName());
+        return ResponseEntity.ok().body(new Response(committeeDetails));
     }
 }