+++ mmms-backend/src/main/resources/data.sql
@@ -8,18 +8,81 @@ INSERT INTO members (
 ) VALUES
       (
           'Anita', 'Shrestha', 'Tribhuvan University', 'Lecturer', 'MSc Physics', 'anita.shrestha@example.com',
-          'admin', '2025-07-01', 'admin', '2025-07-01'
+          'username', '2025-07-01', 'username', '2025-07-01'
       ),
       (
           'Ramesh', 'Khadka', 'Kathmandu University', 'Professor', 'PhD Engineering', 'ramesh.khadka@example.com',
-          'admin', '2025-07-02', 'admin', '2025-07-02'
+          'username', '2025-07-02', 'username', '2025-07-02'
       ),
       (
           'Sita', 'Thapa', 'Pokhara University', 'Coordinator', 'MBA', 'sita.thapa@example.com',
-          'admin', '2025-07-03', 'admin', '2025-07-03'
+          'username', '2025-07-03', 'username', '2025-07-03'
       ),
       (
           'Mel', 'Sitara', 'Pokhara University', 'Coordinator', 'MBA', 'sita.thapa@example.com',
-          'admin', '2025-07-03', 'admin', '2025-07-03'
+          'username', '2025-07-03', 'username', '2025-07-03'
       )
 ;
+
+
+INSERT INTO committees (
+    committee_name,
+    committee_description,
+    created_by,
+    created_date,
+    modified_by,
+    modified_date
+) VALUES (
+             'Finance Committee',
+             'Handles all financial matters of the organization',
+             1,
+             CURDATE(),
+             'username',
+             CURDATE()
+         );
+
+-- get the id of the last inserted entity
+SET @committeeId = LAST_INSERT_ID();
+
+-- Insert a new meeting linked to the newly inserted committee
+INSERT INTO meetings (
+    meeting_name,
+    meeting_description,
+    meeting_held_date,
+    meeting_held_place,
+    created_by,
+    updated_by,
+    created_date,
+    updated_date,
+    committee_id
+) VALUES (
+             'Budget Planning Meeting',
+             'Meeting to discuss the budget allocation for the next fiscal year',
+             '2025-07-08',
+             'Conference Hall A',
+             'username',
+             'username',
+             CURDATE(),
+             CURDATE(),
+             @committeeId
+         );
+
+INSERT INTO decisions (
+    meeting_id,
+    decision
+) VALUES
+      (1, 'Approved FY2025 budget.'),
+      (1, 'Requested audit report by August.');
+
+
+-- Now insert into committee_memberships
+-- Using committee_id = 1 and the newly inserted member_id = 1
+INSERT INTO committee_memberships (
+    committee_id,
+    member_id,
+    role
+) VALUES (
+             1,
+             1,
+             'Chairperson'
+         );